fix(types): Pyright burndown 388->0 across ~65 files (#292)
Resolve all remaining Pyright errors (backend now type-clean). All fixes
behavior-preserving; tests 1379 pass, ruff clean. Only 2 # type: ignore
added (registry.py duck-typed is_premium).

Patterns: None-guards / early-return on Optional access; @overload for
_enrich_event_teams; sentinel/assert for just-created-row invariants;
_LifecycleHost type-only base for the lifecycle mixins' host surface;
lastrowid asserts; EllipsisType sentinel widening; cast for dynamic
getattr/regex results; correct create_channel_service db_factory to a
context-manager factory. Follows #293 (groups.py -144).

+++ teamarr/consumers/matching/racing_matcher.py
@@ -225,7 +225,7 @@ class RacingMatcher:
         stream_norm = normalize_text(ctx.stream_name)
         best_score = 0
         best_event: Event | None = None
-        country_scores: dict[str, int] = {}
+        country_scores: dict[str, float] = {}
 
         for event in events:
             for candidate in (event.name, event.short_name, event.circuit_name):
